requirements.txt
# Dependencias para la automatización de pruebas
selenium==4.15.0  
requests==2.31.0               # Para realizar peticiones HTTP
webdriver-manager==4.0.1
python-dotenv==1.0.0           # Para cargar variables de entorno desde un archivo .env
Faker==22.2.0                 # Para generar datos de prueba aleatorios

# Pytest y dependencias relacionadas
pytest==8.0.0
pytest-xdist==3.3.1            # Para ejecutar pruebas en paralelo                 
pytest-html==4.1.1            # Para reportes en HTML para pytest
requests==2.31.0              # Para pruebas de API
jsonschema==4.23.0            # Para validación de esquemas JSON
PyHamcrest==2.0.4            # Para assertions más expresivas